requirements.txt
scikit-learn>=1.2
numpy>=1.24
pyyaml>=6.0